Account area

+++ LowBank.Api/Models/Account.cs
@@ -2,11 +2,11 @@
 {
     public class Account
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public decimal Amount { get; set; }
         public int Limit { get; set; }
 
-        public Account(int id, decimal amount, int limit)
+        public Account(long id, decimal amount, int limit)
         {
             Id = id;
             Amount = amount;
